server/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-socketio==5.10.0
websockets==11.0.3
aiohttp==3.9.1
pydantic==2.5.2
python-multipart==0.0.7
python-dotenv==1.0.0 